requirements.txt
pandas
plotly
pillow
streamlit
requests
openai>=1.2
python-dotenv
st_pages
requests